fix(marketplace): break circular dependency between commerceState and types
Moved COMMERCE_STATES const and CommerceState type from commerceState.ts
into types.ts, then re-exported them from commerceState.ts so existing
callers (products.ts and marketplace components) keep working unchanged.
Both imports were already type-only, but madge's default parser treats
import-type identically to runtime imports, so the module-level cycle
had to be broken at the graph level.

Verified with: npx madge --circular --extensions ts,svelte ./src
